+++ tests/test_dummy.py
@@ -1,6 +1,0 @@
-"""dummy tests"""
-
-
-def test_dummy() -> None:
-    """Test dummy function."""
-    assert True